Cleaned up the code a LOT.
Moved the logic for building the list of channels to send to the onReady section.

Removed the dependency on Underscore

First, made it so the bot automatically looks through all the server's it's in.
Then it looks for "#annoucements" and checks if it has permissions for it.
If everything is good, it adds it to the broadcast list.

THEN  it goes through the channels in the config file.  If it's in the broadcast list already, it skips it.